feat(Stepper component): updates style classes with theme
