+++ stories/Stepper/Stepper.regression-test.story.jsx
@@ -9,19 +9,19 @@ storiesOf('Stepper', module).add('1 of 6: primary', () => (
       isMobile
       index={3}
       total={8}
-      currentStepText="Salários / benefícios "
+      currentStepText="Salários / benefícios"
       nextStepText="Contrato / Local"
     />
     <Stepper
       index={3}
       total={4}
-      currentStepText="Salários / benefícios "
+      currentStepText="Salários / benefícios"
       nextStepText="Contrato / Local "
     />
     <Stepper
       index={7}
       total={7}
-      currentStepText="Salários / benefícios "
+      currentStepText="Salários / benefícios"
       nextStepText="Contrato / Local "
     />
   </div>
